[refactor] move HTTPResponseHXRedirect
